Changed structure of barChart to make it easier to use (?)
